Move place indexation service to separate service provider
