+++ app/Place/PlaceServiceProvider.php
@@ -4,24 +4,13 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\SearchService\Place;
 
-use CultuurNet\UDB3\Search\ElasticSearch\ElasticSearchDocumentRepository;
-use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\PlaceTransformer;
-use CultuurNet\UDB3\Search\ElasticSearch\PathEndIdUrlParser;
-use CultuurNet\UDB3\Search\JsonDocument\JsonDocumentFetcher;
-use CultuurNet\UDB3\Search\JsonDocument\JsonDocumentTransformer;
-use CultuurNet\UDB3\Search\JsonDocument\JsonTransformerPsrLogger;
-use CultuurNet\UDB3\Search\JsonDocument\TransformingJsonDocumentIndexService;
-use CultuurNet\UDB3\Search\Place\PlaceSearchProjector;
 use CultuurNet\UDB3\SearchService\BaseServiceProvider;
 use CultuurNet\UDB3\SearchService\Offer\OfferSearchControllerFactory;
-use Elasticsearch\Client;
 
 final class PlaceServiceProvider extends BaseServiceProvider
 {
     protected $provides = [
         'place_controller',
-        'place_search_projector',
-        'event_bus_subscribers',
     ];
 
     public function register(): void
@@ -35,48 +24,6 @@ final class PlaceServiceProvider extends BaseServiceProvider
                 return $offerControllerFactory->createFor(
                     $this->parameter('elasticsearch.place.read_index'),
                     $this->parameter('elasticsearch.place.document_type')
-                );
-            }
-        );
-
-        $this->add(
-            'place_search_projector',
-            function () {
-                $service = new TransformingJsonDocumentIndexService(
-                    $this->get(JsonDocumentFetcher::class)->withIncludeMetadata(),
-                    $this->get('place_elasticsearch_transformer'),
-                    $this->get('place_elasticsearch_repository')
-                );
-                $service->setLogger($this->get('logger.amqp.udb3_consumer'));
-
-                return new PlaceSearchProjector($service);
-            },
-            'event_bus_subscribers'
-        );
-
-        $this->add(
-            'place_elasticsearch_transformer',
-            function () {
-                return new JsonDocumentTransformer(
-                    new PlaceTransformer(
-                        new JsonTransformerPsrLogger(
-                            $this->get('logger.amqp.udb3_consumer')
-                        ),
-                        new PathEndIdUrlParser(),
-                        $this->get('offer_region_service')
-                    )
-                );
-            }
-        );
-
-        $this->add(
-            'place_elasticsearch_repository',
-            function () {
-                return new ElasticSearchDocumentRepository(
-                    $this->get(Client::class),
-                    $this->parameter('elasticsearch.place.write_index'),
-                    $this->parameter('elasticsearch.place.document_type'),
-                    $this->get('elasticsearch_indexation_strategy')
                 );
             }
         );
